pyproject.toml
[project]
name = "lektos"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Rust",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
dynamic = ["version"]
dependencies = [
    "python-dotenv>=1.0.1",
    "qdrant-client>=1.12.1",
]
[tool.maturin]
features = ["pyo3/extension-module"]

[tool.uv.workspace]
members = ["python"]
